+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 29/05/2025 - v5.1.2
+
+- (Add) PrusaSlicer printer: Elegoo Saturn 4 (#1028)
+- (Fix) Anycubic ZIP: Incorrect file volume (#983)
+- (Fix) Tool - Change Resolution: In some cases it was clipping content due operation not copying the bounding model but just the absolute center of the image (#1031)
+- (Fix) Tool - Pixel Arithmetic and Stress Tower: A leftover debug method was causing the run in single thread leading to very slow processing
+- (Fix) Prevent a deadloop when calculating the bounding rectangle of all empty layers
+- (Upgrade) .NET from 9.0.4 to 9.0.5
+
 ## 09/05/2025 - v5.1.1
 
 - (Fix) Anycubic ZIP: `System.InvalidOperationException: Sequence contains no elements` when having empty layers (#1023)